Move decluster method to interpolation module
